Move NavigationBar to components
