+++ App/Components/NotificationsBar.js
